Extract ALC and ACC block classes out of block renderer [MAILPOET-2979]
